Move bare router probes into registry
Split the ordered bare-router route probes and option builders out of RouterCommandDefinition into RouterRouteRegistry, leaving the hidden command definition focused on CLI wiring and dispatch.
